+++ lib/TypeOfMap.ts
@@ -1,5 +1,5 @@
 import type { GenericFunction } from "./GenericFunction.js";
-import type { ReadOnlyRecord } from "./ReadOnlyRecord.js";
+import type { NonPrimitive } from "./NonPrimitive.js";
 
 /**
  * `typeof` type map.
@@ -16,8 +16,9 @@ export type TypeOfMap = {
 	readonly bigint: bigint;
 	readonly boolean: boolean;
 	readonly function: GenericFunction;
+	readonly null: null;
 	readonly number: number;
-	readonly object: ReadOnlyRecord;
+	readonly object: NonPrimitive;
 	readonly string: string;
 	readonly symbol: symbol;
 	readonly undefined: undefined;